requirements.txt
fastapi
uvicorn
flask
requests
google-generativeai
load_dotenv
uuid